drivers/device/ringbuffer: move into crazyflie syslink (only remaining user)
