+++ src/drivers/distance_sensor/mb12xx/mb12xx.cpp
@@ -56,7 +56,6 @@
 #include <containers/Array.hpp>
 #include <drivers/device/device.h>
 #include <drivers/device/i2c.h>
-#include <drivers/device/ringbuffer.h>
 #include <drivers/drv_hrt.h>
 #include <drivers/drv_range_finder.h>
 #include <perf/perf_counter.h>
