+++ src/lib/drivers/airspeed/airspeed.cpp
@@ -49,7 +49,6 @@
 
 #include <drivers/drv_airspeed.h>
 #include <drivers/drv_hrt.h>
-#include <drivers/device/ringbuffer.h>
 
 #include <uORB/uORB.h>
 #include <uORB/topics/differential_pressure.h>
