+++ src/lib/drivers/device/CMakeLists.txt
@@ -55,7 +55,6 @@ endif()
 
 px4_add_library(drivers__device
 	CDev.cpp
-	ringbuffer.cpp
 	${SRCS_PLATFORM}
 	)
